+++ src/main/java/com/team6/finalproject/club/enums/JoinTypeEnum.java
@@ -1,4 +1,4 @@
-package com.team6.finalproject.club.entity;
+package com.team6.finalproject.club.enums;
 
 import lombok.Getter;
