+++ src/main/java/com/team6/finalproject/club/interest/entity/InterestMajor.java
@@ -1,22 +1,22 @@
-package com.team6.finalproject.interest.entity;
+package com.team6.finalproject.club.interest.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.util.ArrayList;
+import java.util.List;
+
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor
-public class InterestMinor {
+public class InterestMajor {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String minorName;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "interest_major_id")
-    private InterestMajor interestMajor;
+    @Column(name = "major_name")
+    private String majorName;
 }
