+++ src/main/java/com/team6/finalproject/club/interest/entity/InterestMinor.java
@@ -1,22 +1,22 @@
-package com.team6.finalproject.interest.entity;
+package com.team6.finalproject.club.interest.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.ArrayList;
-import java.util.List;
-
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor
-public class InterestMajor {
+public class InterestMinor {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "major_name")
-    private String majorName;
+    private String minorName;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "interest_major_id")
+    private InterestMajor interestMajor;
 }
