+++ src/main/java/com/team6/finalproject/club/interest/repository/InterestMinorRepository.java
@@ -1,0 +1,7 @@
+package com.team6.finalproject.club.interest.repository;
+
+import com.team6.finalproject.club.interest.entity.InterestMinor;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface InterestMinorRepository extends JpaRepository<InterestMinor, Long> {
+}
