+++ src/main/java/com/team6/finalproject/club/service/ClubService.java
@@ -4,14 +4,12 @@ import com.team6.finalproject.club.dto.ClubRequestDto;
 import com.team6.finalproject.club.dto.ClubResponseDto;
 import com.team6.finalproject.club.dto.InterestMajorDto;
 import com.team6.finalproject.club.dto.InterestMinorDto;
-import com.team6.finalproject.club.entity.ActivityTypeEnum;
+import com.team6.finalproject.club.enums.ActivityTypeEnum;
 import com.team6.finalproject.club.entity.Club;
-import com.team6.finalproject.club.entity.JoinTypeEnum;
+import com.team6.finalproject.club.enums.JoinTypeEnum;
 import com.team6.finalproject.club.repository.ClubRepository;
-import com.team6.finalproject.interest.entity.InterestMajor;
-import com.team6.finalproject.interest.entity.InterestMinor;
-import com.team6.finalproject.interest.repository.InterestMajorRepository;
-import com.team6.finalproject.interest.repository.InterestMinorRepository;
+import com.team6.finalproject.club.interest.entity.InterestMinor;
+import com.team6.finalproject.club.interest.repository.InterestMinorRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
