--- src/main/java/com/team6/finalproject/interest/repository/InterestMinorRepository.java
@@ -1,8 +1,0 @@
-package com.team6.finalproject.interest.repository;
-
-import com.team6.finalproject.interest.entity.InterestMinor;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-public interface InterestMinorRepository extends JpaRepository<InterestMinor, Long> {
-}
